mach-msm: Fix compilation error
FULL_RT config option throws following error:
	smd_private.h: No such file or directory

smd_private.h is present in the same directory, hence change
how it works for non-FULL_RT options but it seems FULL_RT
option affects -I option of the compiler resulting in
compilation error. We should change this we if we find a
better solution or maybe this is how it should have been done.

Change-Id: Iffe35fff1e6ea7c589cfabdf92ab6be0593e89cd
Signed-off-by: Pranav Tipnis <pranav.tipnis@gmail.com>