sound: msm: Fix compilation error
FULL_RT config option throws following error:
        q6voice.h: No such file or directory

q6voice.h is present in the same directory, hence change
how it works for non-FULL_RT options but it seems FULL_RT
option affects -I option of the compiler resulting in
compilation error. We should change this we if we find a
better solution or maybe this is how it should have been done.

Change-Id: Idc8403aa13be28cf6a83a6216f211a2b62f56c07
Signed-off-by: Pranav Tipnis <pranav.tipnis@gmail.com>